+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* **\[core\]** Change the access of `arez.Node.getContext()` from package access to public.
+
 ### [v0.124](https://github.com/arez/arez/tree/v0.124) (2019-01-18)
 [Full Changelog](https://github.com/arez/arez/compare/v0.123...v0.124)